refactor: add useExchangeContext hook as part of ExchangeProvider
