+++ src/providers/ExchangeProvider.tsx
@@ -1,74 +1,11 @@
-import { createContext, PropsWithChildren, useCallback, useEffect, useState } from 'react';
-import { toCeloWei, toStCeloWei } from 'src/formatters/amount';
-import { useAccount } from 'src/hooks/useAccount';
-import { useContracts } from 'src/hooks/useContracts';
-import { Celo, CeloWei, StCelo, StCeloWei } from 'src/types/units';
+import { createContext, PropsWithChildren, useContext } from 'react';
+import { useExchangeRates } from 'src/hooks/useExchangeRates';
 
 interface ExchangeContext {
   celoExchangeRate: number;
   stCeloExchangeRate: number;
   loadExchangeRates: () => Promise<void>;
 }
-
-const useCeloExchangeRate = () => {
-  const { address } = useAccount();
-  const { managerContract } = useContracts();
-
-  const [celoExchangeRate, setCeloExchangeRate] = useState(0);
-
-  const loadCeloExchangeRate = useCallback(async () => {
-    const oneCeloWei = toCeloWei(new Celo('1')).toFixed();
-    const stCeloAmount = new StCeloWei(
-      await managerContract.methods.toStakedCelo(oneCeloWei).call({ from: address })
-    );
-    setCeloExchangeRate(stCeloAmount.dividedBy(oneCeloWei).toNumber());
-  }, [managerContract, address]);
-
-  return {
-    celoExchangeRate,
-    loadCeloExchangeRate,
-  };
-};
-
-const useStCeloExchangeRate = () => {
-  const { address } = useAccount();
-  const { managerContract } = useContracts();
-
-  const [stCeloExchangeRate, setStCeloExchangeRate] = useState(0);
-
-  const loadStCeloExchangeRate = useCallback(async () => {
-    const oneStCeloWei = toStCeloWei(new StCelo('1')).toFixed();
-    const celoAmount = new CeloWei(
-      await managerContract.methods.toCelo(oneStCeloWei).call({ from: address })
-    );
-    setStCeloExchangeRate(celoAmount.dividedBy(oneStCeloWei).toNumber());
-  }, [managerContract, address]);
-
-  return {
-    stCeloExchangeRate,
-    loadStCeloExchangeRate,
-  };
-};
-
-const useExchangeRate = () => {
-  const { celoExchangeRate, loadCeloExchangeRate } = useCeloExchangeRate();
-  const { stCeloExchangeRate, loadStCeloExchangeRate } = useStCeloExchangeRate();
-
-  const loadExchangeRates = useCallback(async () => {
-    await Promise.all([loadCeloExchangeRate(), loadStCeloExchangeRate()]);
-  }, [loadCeloExchangeRate, loadStCeloExchangeRate]);
-
-  useEffect(() => {
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    loadExchangeRates();
-  }, [loadExchangeRates]);
-
-  return {
-    celoExchangeRate,
-    stCeloExchangeRate,
-    loadExchangeRates,
-  };
-};
 
 export const ExchangeContext = createContext<ExchangeContext>({
   celoExchangeRate: 0,
@@ -77,7 +14,7 @@ export const ExchangeContext = createContext<ExchangeContext>({
 });
 
 export const ExchangeProvider = ({ children }: PropsWithChildren) => {
-  const { celoExchangeRate, stCeloExchangeRate, loadExchangeRates } = useExchangeRate();
+  const { celoExchangeRate, stCeloExchangeRate, loadExchangeRates } = useExchangeRates();
 
   return (
     <ExchangeContext.Provider
@@ -91,3 +28,12 @@ export const ExchangeProvider = ({ children }: PropsWithChildren) => {
     </ExchangeContext.Provider>
   );
 };
+
+export function useExchangeContext() {
+  const { celoExchangeRate, stCeloExchangeRate } = useContext(ExchangeContext);
+
+  return {
+    celoExchangeRate,
+    stCeloExchangeRate,
+  };
+}
